finance_tracker/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.7
boto3==1.28.39